control/requirements.txt
python-dotenv
paho-mqtt==1.6.1
